feat: knowledge tab add button (#1277)
* 🚧 🌐 knowledge traductions

* 🔥 replace useKnowledge hook

* ✨ AddKnowledge in knowledge  tab

* 🚧 extract feedBrain in chat hook

* ♻️ use feedBrain method as a prop in KnowledgeToFeedInput

* ✨ custom feedBrain for chat and brain management

* ✨ add loader when deleting

* 🚑 remove chat after auto creation for each feed action

* 🚚 rename KnowledgeProvider into KnowledgeToFeedProvider

+++ frontend/app/brains-management/[brainId]/components/BrainManagementTabs/components/KnowledgeTab/KnowledgeItem/DeleteKnowledge.tsx
@@ -1,12 +1,12 @@
 "use client";
 
+import { AiOutlineLoading3Quarters } from "react-icons/ai";
 import { MdDelete } from "react-icons/md";
 
 import { useBrainContext } from "@/lib/context/BrainProvider/hooks/useBrainContext";
 import { Knowledge } from "@/lib/types/Knowledge";
 
 import { useKnowledgeItem } from "./useKnowledgeItem";
-
 export const DeleteKnowledge = ({
   knowledge,
 }: {
@@ -18,18 +18,18 @@ export const DeleteKnowledge = ({
 
   const canDeleteFile = currentBrain?.role === "Owner";
 
-  console.log("isDeleting", isDeleting);
+  if (!canDeleteFile) {
+    return <></>;
+  }
 
-  return (
-    <>
-      {canDeleteFile && (
-        <button
-          className="text-red-600 hover:text-red-900"
-          onClick={() => void onDeleteKnowledge(knowledge)}
-        >
-          <MdDelete size="20" />
-        </button>
-      )}
-    </>
+  return isDeleting ? (
+    <AiOutlineLoading3Quarters />
+  ) : (
+    <button
+      className="text-red-600 hover:text-red-900"
+      onClick={() => void onDeleteKnowledge(knowledge)}
+    >
+      <MdDelete size="20" />
+    </button>
   );
 };
